requirements.txt
beautifulsoup4
mysql-connector-python
python-dotenv
pandas
schedule
matplotlib
numpy
seaborn